requirements.txt
Flask==1.1.1
line_bot_sdk==1.16.0
numpy==1.16.4
Pillow==6.1.0
tensorflow==2.0
keras==2.1.2



